refactor(classify): de-duplicate LLM JSON extraction + base64/MIME helpers [#130]
Hoist the copy-pasted markdown-fence JSON extraction and data-URL
stripping logic into shared modules:

- New src/lib/classify/json.ts exports extractJsonObject<T>(), the
  fence-strip + outermost {…} slice + JSON.parse used by both
  parseClassificationResponse (types.ts) and parseObservation
  (observe.ts). Type-specific validation stays in the parsers.
- New src/lib/classify/image-utils.ts exports stripDataUrlPrefix()
  (the guarded variant from preprocess.ts); anthropic/google providers
  and preprocess now import it instead of local copies.

extractMediaType and extractMimeType are intentionally left untouched
(distinct signatures/semantics). No behavior change.
