+++ nexa/src/lib/classify/types.ts
@@ -1,3 +1,5 @@
+import { extractJsonObject } from "./json";
+
 export const ISSUE_TYPES = [
   "ROAD_DAMAGE",
   "STREETLIGHT_OUTAGE",
@@ -99,15 +101,5 @@ export function buildClassificationPrompt(
 
 /** Parse model JSON even when wrapped in markdown fences or extra prose. */
 export function parseClassificationResponse(raw: string): ClassificationResult {
-  let text = raw
-    .trim()
-    .replace(/```(?:json)?\s*/gi, "")
-    .replace(/```/g, "");
-  text = text.trim();
-  const start = text.indexOf("{");
-  const end = text.lastIndexOf("}");
-  if (start === -1 || end === -1 || end <= start) {
-    throw new SyntaxError("No JSON object in model response");
-  }
-  return JSON.parse(text.slice(start, end + 1)) as ClassificationResult;
+  return extractJsonObject<ClassificationResult>(raw);
 }
